Move plugin extension tests to a new class
